% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{import}
\usage{
import(file, ...)
}
\arguments{
\item{file}{A path to the file to be imported into the R Global Environment}

\item{...}{Additional argument used by the import function, optional}
}
\value{
A data frame or a tibble depending on the file's extension.
}
\description{
Import a data frame from a selected file
}
\details{
Extensions supported include .sas7bdat, .dta, .sav, .xlsx, .xls, and most delimited files
}
\examples{
\dontrun{ df <- import(file) }

}
