% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use}
\alias{use}
\title{use}
\usage{
use(pkg)
}
\arguments{
\item{pkg}{A R package available on CRAN}
}
\value{
This function will only return an error message when failing, otherwise not output will be returned
}
\description{
Load a package
}
\details{
Will also download the package if it has not been previously downloaded by the user
}
\examples{
\dontrun{ use(package_name) }

}
