% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import Function}
\usage{
import(file, ...)
}
\arguments{
\item{file}{a character string of the file.}

\item{...}{Additional arguments passed to the underlying import functions.
For example, this can control column classes for delimited file types,
or control the use of haven for Stata and SPSS or readxl for Excel (.xlsx) format.}
}
\value{
return a dataframe.
}
\description{
\code{import} reads in a dataframe from a file.
}
\details{
This function imports a data frame from a data file using a filename, or prompt
the user to choose a file without the need to provide a filename.
}
\examples{
\dontrun{df <- import("datasetexample.xlsx")}
\dontrun{df <- import("datasetexample.xlsx", col_names=TRUE)}
}
\seealso{
\href{https://www.rdocumentation.org/packages/haven/versions/2.3.1/topics/read_sas}{read_sas}, \href{https://www.rdocumentation.org/packages/haven/versions/2.3.1/topics/read_dta}{read_stata}, \href{https://www.rdocumentation.org/packages/haven/versions/2.3.1/topics/read_spss}{read_spss}, \href{https://www.rdocumentation.org/packages/readxl/versions/1.3.1/topics/read_excel}{read_excel}, \href{https://www.rdocumentation.org/packages/vroom/versions/1.0.2/topics/vroom}{vroom}.
}
\author{
Ahmed Aldarmaki \href{mailto:aaldarmaki@wesleyan.edu}{aaldarmaki@wesleyan.edu}
}
