% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import Function}
\usage{
import(file, ...)
}
\arguments{
\item{file}{a character string naming a file.}

\item{...}{Additional arguments passed to the underlying import functions.
For example, this can control column classes for delimited file types,
or control the use of haven for Stata and SPSS or readxl for Excel (.xlsx) format.}
}
\value{
return a dataframe.
}
\description{
\code{import} reads in a dataframe from a file.
}
\details{
This function imports a data frame or matrix from a data file
with the file format based on the file extension, and it also prompt
user to choose a file without the need to provide a filename.
}
\examples{
\dontrun{df <- import("datasetexample.xlsx")}
}
\seealso{
\code{\link{read_sas}}, \code{\link{read_stata}}, \code{\link{read_spss}}, \code{\link{read_excel}}, \code{\link{vroom}}.
}
\author{
Ahmed Aldarmaki \href{mailto:aaldarmaki@wesleyan.edu}{aaldarmaki@wesleyan.edu}
}
