% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import new data sets}
\usage{
import(file, ...)
}
\arguments{
\item{file}{name of file}

\item{...}{extra specification for file import}
}
\value{
a data set in the environment
}
\description{
Imports files into R, can read in multiple file types.
}
\examples{
import("C:/Users/ascol/Downloads/testdata.csv")

}
