% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import new data sets}
\usage{
import(file, ...)
}
\arguments{
\item{file}{name of file}

\item{...}{extra specification for file import}
}
\value{
a data set in the environment
}
\description{
Imports data sets into R
}
\details{
Can import files of multiple type into the global environment.
}
\examples{
\dontrun{ import("C:/Users/ascol/Downloads/testdata.csv")}

}
\seealso{
\code{\link{haven}}
}
