% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import Function}
\usage{
import(file, ...)
}
\arguments{
\item{file}{a file containing data the user wants to import into R}

\item{...}{an argument passed to the import function}
}
\value{
a data frame
}
\description{
Import function takes a file specified by the user and imports it into R as a data frame.
}
\examples{
\dontrun{
mydata <- import(datafile.dta)
}
}
\seealso{
\code{\link[=read.csv]{read.csv()}} \code{\link[readxl:read_excel]{readxl::read_excel()}}
}
