% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import}
\usage{
import(file, ...)
}
\arguments{
\item{file}{data from an external source}

\item{...}{other features that user may want to include}
}
\value{
a dataset
}
\description{
\code{import} imports a dataset from external
files and returns that dataset
}
\examples{
# newdata <- import(mydata.spss)
}
