% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import A Dataframe}
\usage{
import(file, ...)
}
\arguments{
\item{file}{path to a file.}

\item{...}{additional arguments passed to the function.}
}
\value{
a dataframe
}
\description{
import will import a dataframe from a variety of file formats.
}
\examples{
\dontrun{
import("mtcars.xls")
import("mtcars.dta")
}
}
\seealso{
\code{\link[=read_sas]{read_sas()}},\code{\link[=read_stata]{read_stata()}},\code{\link[=read_spss]{read_spss()}},\code{\link[=read_excel]{read_excel()}},\code{\link[=vroom]{vroom()}}
}
