% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{import}
\usage{
import(file, ...)
}
\arguments{
\item{file}{An optional character string naming a path to a file. By default, the user will be prompted to choose a file.}

\item{...}{additional arguments passed to the function.}
}
\value{
A Dataframe
}
\description{
Imports a data frame from a data file with the file format based on the file extension
}
\details{
This function imports a data frame from a data file with the file format based on the file extension.
\code{import} supports a variety of file extensions including SAS (.sas7bdat), Stata (.dta), SPSS (.sav), Excel (.xls and .xlsx), Comma-separated data (.csv), and Tab-separated data (.tsv).
If the file path is not specified, \code{import} will prompt the user to choose a file.
}
\examples{
\dontrun{
# import XLS or SAV as a `data.frame`
df = import("mtcars.xls")
df = import("mtcars.sav")

# pass arguments to underlying import function
df = import("mtcars.dta",nmax = 5)
}
}
\seealso{
\href{https://www.rdocumentation.org/packages/haven/versions/2.3.1/topics/read_spss}{\code{read_spss}},\href{https://www.rdocumentation.org/packages/readxl/versions/1.3.1/topics/read_excel}{\code{read_excel}},\href{https://www.rdocumentation.org/packages/vroom/versions/1.0.2/topics/vroom}{\code{vroom}},\href{https://www.rdocumentation.org/packages/haven/versions/2.3.1/topics/read_dta}{\code{read_sas}}
}
