% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import}
\usage{
import(file, ...)
}
\arguments{
\item{file}{name of file in path.}

\item{...}{extra parameters passed into files.}
}
\value{
returns a data frame of file.
}
\description{
Imports a variety of different data files
}
\examples{
# import(data.csv)
}
\seealso{
\itemize{
\item \code{\link[haven:read_sas]{haven::read_sas()}} read sas7bdat files
\item \code{\link[haven:read_dta]{haven::read_stata()}} read dta files
\item \code{\link[haven:read_spss]{haven::read_spss()}} read sav files
\item \code{\link[readxl:read_excel]{readxl::read_excel()}} read excel files
\item \code{\link[vroom:vroom]{vroom::vroom()}} read most files fast
}
}
