% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import}
\usage{
import(file, ...)
}
\arguments{
\item{file}{name of file in path.}

\item{...}{extra parameters passed into files.}
}
\value{
returns a data frame of file.
}
\description{
Imports a variety of different data files
}
\examples{
# import(data.csv)
}
