% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import}
\usage{
import(file, ...)
}
\arguments{
\item{file}{name of file in path.}

\item{...}{passes additional parameters into file options.}
}
\value{
dataframe object
}
\description{
This function imports a variety of different data files.
}
\examples{
\dontrun{import(data.csv)}
}
\seealso{
\itemize{
\item \code{\link[haven:read_sas]{haven::read_sas()}} read sas files
\item \code{\link[haven:read_dta]{haven::read_stata()}} read stata files
\item \code{\link[haven:read_spss]{haven::read_spss()}} read spss files
\item \code{\link[readxl:read_excel]{readxl::read_excel()}} read excel files
\item \code{\link[vroom:vroom]{vroom::vroom()}} reads files quicker
}
}
