% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Make Project}
\usage{
make_project(
  project,
  path = getwd(),
  folders = c("figures", "documentation", "data", "reports", "R"),
  readme = "README.md",
  git = FALSE
)
}
\arguments{
\item{project}{the name of the project, given as a name or character string.}

\item{path}{the path of where the project should be created, the default is the current working directory.}

\item{folders}{a character vector of the folders created in the project.}

\item{readme}{creates a README text file that introduces and explains the project being created.}

\item{git}{a boolean value indicating whether to initialise a Git repository
and add important files to .gitignore.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will create all of the scaffolding for a new project.
It will set up all of the relevant directories and their initial contents.
}
\details{

}
\examples{
# ADD_EXAMPLES_HERE
}
