% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Make Project Function}
\usage{
make_project(
  project,
  path = getwd(),
  folders = c("figures", "documentation", "data", "reports", "R"),
  readme = "README.md",
  git = FALSE
)
}
\arguments{
\item{project}{name of project}

\item{path}{location project will be created in}

\item{folders}{names of the folders in the project}

\item{readme}{name of readme file created in project}

\item{git}{version control}
}
\description{
The make project function creates a new project for the user to work in.
The function also creates a set of sub-folders stored in the same working directory.
}
\examples{
\dontrun{ make_project(project = myproject)}
\dontrun{ make_project(project = myproject2, folders = c("first", "second", "third", "fourth", "R"))}
}
