% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Make Project}
\usage{
make_project(
  project,
  path = getwd(),
  folders = c("figures", "documentation", "data", "reports", "R"),
  readme = "README.md",
  git = FALSE
)
}
\arguments{
\item{project}{The name of the project}

\item{path}{this is the path where the user wants the project to be located. The default will be the current working directory}

\item{folders}{the names of the folders the files from the project will go in. The default are "figures", "documentation","data", "reports", "R"}

\item{readme}{the file where the description of the file will go. The default is "README.md"}

\item{git}{is defaulted at FALSE}
}
\value{
RETURN_DESCRIPTION
}
\description{
This function creates a package which makes path directories, a READ.me file and creates
the RStudio Project
}
\examples{
# myproject <- make_project(newproject)
}
