% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{create a project}
\usage{
make_project(
  project,
  path = getwd(),
  folders = c("figures", "documentation", "data", "reports", "R"),
  readme = "README.md",
  git = FALSE
)
}
\arguments{
\item{project}{A character string containing the name for this new project.}

\item{path}{An optional character string containing the path for this new project. By default, the path is set to the current directory.}

\item{folders}{An optional character vector containing the names of the subdirectories. By default, the project will contain only five subdirectories with default names.}

\item{readme}{An optional character string containing the name of the readme file. By default, the name of the readme file will be "README.md".}

\item{git}{An optional boolean value indicating whether the project files should be committed to Github (\code{TRUE}) or not (\code{FALSE}). Default is \code{FALSE}.}
}
\value{
No value is returned (this function is called for its side effects.)
}
\description{
Creates a new project and sets up all of the relevant directories and their initial contents.
}
\details{
This function creates a new project and sets up all of the relevant directories and their initial contents. If the path of
the project is not specified, the path will be set to the current directory. The default behavior is to add 5 essential
subdirectories, but it is also possible to specify or add your own subdirectories. \code{make_project} also creates
a readme file called "README.md" by default. You can specify the name of the readme file. The project files
will be commited to Github if \code{git = TRUE}, however, the default is \code{FALSE}.
}
\examples{
\dontrun{
make_project("newproject")

# provide a custom path for the project.
make_project("newproject", path = "path/to/new-project")

# automatically commit the new project files to Github

make_project("secondproject", git = TRUE)
}
}
