% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Create A New Project}
\usage{
make_project(
  project,
  path = getwd(),
  folders = c("figures", "documentation", "data", "reports", "R"),
  readme = "README.md",
  git = FALSE
)
}
\arguments{
\item{project}{A character vector containing the name for this new project.}

\item{path}{Set the path of the new project to the current directory.}

\item{folders}{Set the names of the subdirectories in the new project.}

\item{readme}{Set the name and format of the readme file in the new project.}

\item{git}{Add git if true.}
}
\value{
a project
}
\description{
make_project will create a new project
}
\examples{
\dontrun{
make_project("firstproject")
make_project("secondproject", git = TRUE)
}
}
