% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Make a Package}
\usage{
make_project(
  project,
  path = getwd(),
  folders = c("figures", "documentation", "data", "reports", "R"),
  readme = "README.md",
  git = FALSE
)
}
\arguments{
\item{project}{name of project.}

\item{path}{location of project.}

\item{folders}{names of relevant folders.}

\item{readme}{readme. Default is markdown file.}

\item{git}{enable git. Default is FALSE.}
}
\value{
a RProject.
}
\description{
This function creates a project and relevant files within the project location.
}
\examples{
# make_project(cats)
}
