% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Make Project}
\usage{
make_project(
  project,
  path = getwd(),
  folders = c("figures", "documentation", "data", "reports", "R"),
  readme = "README.md",
  git = FALSE
)
}
\arguments{
\item{project}{name of project.}

\item{path}{location of project.}

\item{folders}{names of relevant folders to create (figures, documentation, data, reports, R)}

\item{readme}{readme, default is a markdown file.}

\item{git}{enables git, default is false.}
}
\description{
This function creates an R project with relevant folders in project location.
}
\examples{
\dontrun{make_project(cats)}
}
