% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use}
\alias{use}
\title{Loading/Attaching of Packages}
\usage{
use(pkg)
}
\arguments{
\item{pkg}{the name of a package, given as a name or literal character string.}
}
\description{
\code{use} loads and attaches add-on packages.
}
\details{
This function loads the namespace of the package with name \code{pkg} and attach it on the search list.
If package is not installed, it asks if you want to install it, and if yes loads/attaches the package.
Else, it returns a message.If the package does not exists, it returns FALSE and gives a warning. It checks
and updates the list of currently attached packages and do not reload a namespace which is already loaded.
}
\examples{
\dontrun{use(dplyr)}
\dontrun{use("dplyr")}
}
\author{
Ahmed Aldarmaki \href{mailto:aaldarmaki@wesleyan.edu}{aaldarmaki@wesleyan.edu}
}
