% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use}
\alias{use}
\title{Install new packages}
\usage{
use(pkg)
}
\arguments{
\item{pkg}{package name}
}
\description{
Installs new packages to be used into the project.
}
\details{
Similar to library but if the package is not yet loaded will ask user whether or not they want to load the package.
}
\examples{
\dontrun{ use(ggplot2)}
}
