% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use}
\alias{use}
\title{Use Package Function}
\usage{
use(pkg)
}
\arguments{
\item{pkg}{package to install and load}
}
\description{
The use (package) function installs and loads a package specified by the user.
}
\examples{
# CHECK_EXAMPLES_HERE
\dontrun{
use(dplyr)
use(ggplot2)
}
}
\seealso{
\code{\link[=install.packages]{install.packages()}} \code{\link[=library]{library()}} \code{\link[=require]{require()}}
}
