% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use}
\alias{use}
\title{Use}
\usage{
use(pkg)
}
\arguments{
\item{pkg}{a package}
}
\value{
Will return either a loaded package or will not load a package based on user input.
}
\description{
This function asks the user if they want to download a package and if so  will download it.
}
\examples{
# use(my_package)
}
