% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use}
\alias{use}
\title{Use a package}
\usage{
use(pkg)
}
\arguments{
\item{pkg}{A vector containing the name for a package.}
}
\value{
the input object is returned silently
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\dontrun{
use(dplyr)
use(ggplot2)
}
}
