% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use}
\alias{use}
\title{Installing/Attaching an add-on package}
\usage{
use(pkg)
}
\arguments{
\item{pkg}{the name of a package given as a name or character string.}
}
\value{
returns (invisibly) a logical indicating whether
the required package is available
}
\description{
use() installs and attaches add-on packages
}
\details{
use() checks and updates the list of currently
attached packages and do not reload a namespace which
is already loaded. If not already installed,
the function asks the user whether they want to install the package.
}
\examples{
\dontrun{use(vroom)}
\dontrun{use("ggplot2")}
}
