% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use}
\alias{use}
\title{Installing and Attaching Packages}
\usage{
use(pkg)
}
\arguments{
\item{pkg}{a character string of the name of the package or just the name of
the package}
}
\value{
no value is returned
}
\description{
\code{use} installs a package or attach add-on packages
}
\details{
This function will load the namespace of the package and attach it
to the search list. If the package is not install, it will prompt a
a dialogue for user to confirm the installation.
}
\examples{
# use("usethis")
# use(usethis)
}
\author{
Cher Qin \href{mailto:qqin@wesleyan.edu}{qqin@wesleyan.edu}
}
